@@ -1,13 +1,15 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 31
     defaultConfig {
         applicationId "it.m2app.screamreceiver"
-        minSdkVersion 22
-        targetSdkVersion 26
+        minSdkVersion 28
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
+        vectorDrawables.useSupportLibrary = true
     }
     buildTypes {
         release {
@@ -19,5 +21,16 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation('androidx.appcompat:appcompat:1.4.1')
+    implementation("androidx.core:core-ktx:1.7.0")
+    implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version")
+    implementation('org.jetbrains.kotlinx:kotlinx-coroutines-core:1.5.0')
+    implementation('io.github.microutils:kotlin-logging-jvm:2.1.20')
+    implementation('androidx.lifecycle:lifecycle-runtime-ktx:2.4.1')
+    implementation('hu.autsoft:krate:2.0.0')
+    implementation('com.frybits.harmony:harmony:1.1.11')
+    implementation('com.github.anastr:speedviewlib:1.6.0')
+}
+repositories {
+    mavenCentral()
 }
